Refactoring: name private members with "_" to make code cleaner
According to "PEP 8" single leading underscore is a weak "internal use"
indicator. It just shows other programmers not to use it.

Double leading underscores result in name mangling. Though it is still
possible to use susch variable with "_ClassName__MemberName". So there is no
need to polute the code with lots of "__"

References:
[1] PEP 8 -- Style Guide for Python Code, Descriptive: Naming Styles
[2] Dan Bader, The Meaning of Underscores in Python